damn still much to do (fill!)

+++ Isolines/src/ru/nsu/fit/g16201/migranov/model/Model.java
@@ -93,11 +93,11 @@ public class Model {
         return maxValue;
     }
 
-    public double getM() {
+    public int getM() {
         return m;
     }
 
-    public double getK() {
+    public int getK() {
         return k;
     }
